Configured auth handler in all routes of tasks

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTaskFilterDto } from './dto/get-task-filter.dto';
 import { TaskStatus } from './task-status.enum';
@@ -14,8 +15,11 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) { }
 
-    async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(
+        filterDto: GetTaskFilterDto,
+        user: User
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
     // getTaskWithFilters(filterDto: GetTaskFilterDto): Task[] {
@@ -37,8 +41,8 @@ export class TasksService {
     //     return tasks;
     // }
 
-    async getTaskById(id: number) {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User) {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
         if (!found) {
             throw new NotFoundException(`Task with id "${id}" not found`);
@@ -47,20 +51,20 @@ export class TasksService {
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
 
-        return this.taskRepository.createTask(createTaskDto);
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
         if (result.affected === 0) {
             throw new NotFoundException(`Task with id "${id}" not found`);
         }
     }
 
-    async updateTask(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTask(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
